+++ modules/@apostrophecms/schema/lib/addFieldTypes.js
@@ -95,9 +95,8 @@ module.exports = (self) => {
 
   self.addFieldType({
     name: 'string',
-    convert: function (req, field, data, destination) {
+    convert(req, field, data, destination) {
       destination[field.name] = self.apos.launder.string(data[field.name], field.def);
-
       destination[field.name] = checkStringLength(destination[field.name], field.min, field.max);
       // If field is required but empty (and client side didn't catch that)
       // This is new and until now if JS client side failed, then it would
@@ -105,8 +104,16 @@ module.exports = (self) => {
       if (field.required && (_.isUndefined(data[field.name]) || !data[field.name].toString().length)) {
         throw self.apos.error('required');
       }
+
+      if (field.pattern) {
+        const regex = new RegExp(field.pattern);
+
+        if (!regex.test(destination[field.name])) {
+          throw self.apos.error('invalid');
+        }
+      }
     },
-    index: function (value, field, texts) {
+    index(value, field, texts) {
       const silent = field.silent === undefined ? true : field.silent;
       texts.push({
         weight: field.weight || 15,
@@ -114,10 +121,22 @@ module.exports = (self) => {
         silent: silent
       });
     },
-    isEmpty: function (field, value) {
+    isEmpty(field, value) {
       return !value.length;
     },
-    addQueryBuilder: function (field, query) {
+    validate(field, options, warn, fail) {
+      if (!field.pattern) {
+        return;
+      }
+
+      const isRegexInstance = field.pattern instanceof RegExp;
+      if (!isRegexInstance && typeof field.pattern !== 'string') {
+        fail('The pattern property must be a RegExp or a String');
+      }
+
+      field.pattern = isRegexInstance ? field.pattern.source : field.pattern;
+    },
+    addQueryBuilder(field, query) {
       query.addBuilder(field.name, {
         finalize: function () {
           if (self.queryBuilderInterested(query, field.name)) {